+++ src/components/analyser/search/DashboardCards.tsx
@@ -1,60 +1,11 @@
 import React from 'react';
-import { navigate } from '@reach/router';
 import InfiniteScroll from 'react-infinite-scroll-component';
-import {
-  Card,
-  CardContent,
-  CardMedia,
-  Grid,
-  Divider,
-  makeStyles,
-  GridList,
-  ButtonBase,
-  useTheme,
-} from '@material-ui/core';
+import { Grid, makeStyles, GridList } from '@material-ui/core';
 import CircularProgress from '@material-ui/core/CircularProgress';
-import { useTranslation } from 'react-i18next';
 import * as API from '../../../analyser/APIClient';
-import TextOverText from '../../shared/TextOverText';
-import CompanyLogo from '../CompanyLogo';
-import { convertPercentToColor } from '../../../analyser/Helper';
+import { DashboardCardsRow } from './DashboardCardRow';
 
 const useStyles = makeStyles(() => ({
-  card: {
-    width: 289,
-    margin: 10,
-    transition: '0.3s',
-    '&:hover': {
-      boxShadow: '0 16px 70px -12.125px rgba(0,0,0,0.3)',
-    },
-  },
-  media: {
-    height: 200,
-    width: 289,
-    placeItems: 'center',
-    alignItems: 'center',
-    objectFit: 'scale-down',
-    padding: 20,
-  },
-  content: {
-    margin: 10,
-  },
-  divider: {
-    margin: `${30}px 0`,
-  },
-  cardAction: {
-    display: 'block',
-    textAlign: 'initial',
-  },
-  rightBound: {
-    float: 'right',
-  },
-  leftBound: {
-    float: 'left',
-  },
-  paddingBottom: {
-    paddingBottom: 40,
-  },
   grid: {
     overflow: 'hidden' /* Hide scrollbars */,
   },
@@ -63,68 +14,16 @@ const useStyles = makeStyles(() => ({
   },
 }));
 
-export type DashboardCardsRowProps = {
-  stock: API.Stock;
-};
-
-/**
- * This component displays stock and some additional information using cards
- */
-export const DashboardCardsRow: React.FC<DashboardCardsRowProps> = ({
-  stock,
-}) => {
-  const classes = useStyles();
-  const { t } = useTranslation();
-  const theme = useTheme();
-
-  return (
-    <Card className={classes.card}>
-      <ButtonBase
-        className={classes.cardAction}
-        onClick={() => navigate(`analyser/${stock.symbol}`)}
-      >
-        <CardMedia>
-          <CompanyLogo stockOverview={stock} style={classes.media} />
-        </CardMedia>
-        <CardContent className={classes.content}>
-          <TextOverText
-            top={`${stock.symbol}`}
-            bottom={`${stock.name}`}
-            colorTop={theme.palette.grey[700]}
-            sizeBottom="1.3rem"
-            alignment="left"
-          />
-          <Divider className={classes.divider} light />
-          <div className={classes.paddingBottom}>
-            <div className={classes.leftBound}>
-              <TextOverText
-                top={t('analyser.dashboard.view.cards.lastPrice')}
-                colorTop={theme.palette.grey[700]}
-                bottom={`${stock.price}`}
-                currency={`${stock.currency}`}
-                sizeBottom="1.3rem"
-              />
-            </div>
-            <div className={classes.rightBound}>
-              <TextOverText
-                top={t('stock.30d')}
-                bottom={`${stock.per30d.toFixed(2)}%`}
-                colorTop={theme.palette.grey[700]}
-                colorBottom={convertPercentToColor(stock.per30d)}
-                sizeBottom="1.3rem"
-              />
-            </div>
-          </div>
-        </CardContent>
-      </ButtonBase>
-    </Card>
-  );
-};
-
 export type DashboardCardsProps = {
   stocks: API.Stock[];
 };
 
+/**
+ * used to display a grid of cards to select stocks from
+ *
+ * @param stocks list of stocks to get stock overview from
+ *
+ */
 const DashboardCards: React.FC<DashboardCardsProps> = ({ stocks }) => {
   const classes = useStyles();
 
@@ -144,7 +43,7 @@ const DashboardCards: React.FC<DashboardCardsProps> = ({ stocks }) => {
     setHasMore(true);
     // a fake async api call like which sends
     // 5 more stocks in 1.5 secs
-    // TODO replace with async API call
+    // TODO replace with async API call if more stocks are provided
     setTimeout(() => {
       const newItems = items.concat(
         stocks.slice(items.length, items.length + 5)
