+++ src/components/analyser/search/DashboardTable.tsx
@@ -1,4 +1,4 @@
-// Initally based on Portfolio's DashboardTable.tsx
+// Initially based on Portfolio's DashboardTable.tsx
 
 import React from 'react';
 import InfiniteScroll from 'react-infinite-scroll-component';
@@ -34,13 +34,14 @@ const useStyles = makeStyles(() =>
   })
 );
 
+// used for sorting
 type Order = 'asc' | 'desc';
 
 export type DashboardTableProps = {
   stocks: API.Stock[];
 };
 
-// desc sort comparator
+/**  desc sort comparator */
 function descendingComparator<T>(a: T, b: T, orderBy: keyof T) {
   if (b[orderBy] < a[orderBy]) {
     return -1;
@@ -50,7 +51,12 @@ function descendingComparator<T>(a: T, b: T, orderBy: keyof T) {
   }
   return 0;
 }
-// function: take asc or desc order and the property to sort the stocks
+
+/** function: take asc or desc order and the property to sort the stocks
+ * @param items items to sort
+ * @param Order current order
+ * @param orderBy key to order by
+ */
 function sortStocks(
   items: API.Stock[],
   order: Order,
@@ -61,10 +67,16 @@ function sortStocks(
     : items.sort((a, b) => -descendingComparator(a, b, orderBy));
 }
 
+/**
+ * Table to display stock data. Can be sorted and has infinity loading implemented
+ *
+ * @param stocks list of stocks to get stock overview from
+ *
+ */
 const DashboardTable: React.FC<DashboardTableProps> = ({ stocks }) => {
   const classes = useStyles();
 
-  // currently dispayled stocks in lazy loading
+  // currently displayed stocks in lazy loading
   const [items, setItems] = React.useState<API.Stock[]>(stocks.slice(0, 10));
   // boolean if more stocks are available for loading
   const [hasMore, setHasMore] = React.useState<boolean>(true);
@@ -86,7 +98,7 @@ const DashboardTable: React.FC<DashboardTableProps> = ({ stocks }) => {
     setHasMore(true);
     // a fake async api call like which sends
     // 5 more stocks in 1.5 secs
-    // TODO replace with async API call
+    // TODO replace with async API call if more stocks are available
     setTimeout(() => {
       const newItems = items.concat(
         sortedStocks.slice(items.length, items.length + 5)
@@ -102,7 +114,7 @@ const DashboardTable: React.FC<DashboardTableProps> = ({ stocks }) => {
     setOrderByKey(property);
   };
 
-  // update sorted stocks if new sotck, new items, new order or new orderByKey
+  // update sorted stocks if new stock, new items, new order or new orderByKey
   React.useEffect(() => {
     setSortedStocks(sortStocks(stocks, order, orderByKey));
     setItems(sortedStocks.slice(0, 10));
