+++ src/components/analyser/search/DashboardTableRow.tsx
@@ -39,12 +39,14 @@ const useStyles = makeStyles(({ palette }: Theme) =>
   })
 );
 
-export type DashboardTableRowProps = {
+type DashboardTableRowProps = {
   stock: API.Stock;
 };
 
 /**
  * This component displays the actual data of individual stocks in the rows of the table
+ *
+ * @param stock single stock overview to display in a row
  */
 const DashboardTableRow: React.FC<DashboardTableRowProps> = ({ stock }) => {
   const [hover, setHover] = React.useState<boolean>(false);
