+++ src/components/analyser/search/SearchOption.tsx
@@ -24,7 +24,7 @@ const useStyles = makeStyles((theme: Theme) =>
 /**
  * A option will show the stock's name and will navigate to its detail page when it has been clicked.
  *
- * @param stock - Stock which related to a singel option
+ * @param stock - Stock which related to a single option
  * @return A single option in the pop up list of search bar
  */
 const SearchOption: React.FC<SearchOptionProps> = ({ stock }) => {
